requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
albucore==0.0.17
albumentations==1.4.18
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
builtin-interfaces==1.2.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1725278078093/work/certifi
charset-normalizer==3.4.4
composition-interfaces==1.2.2
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1695554215751/work
cv-bridge==3.2.1
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
domain-coordinator==0.10.0
einops==0.8.1
eval_type_backport==0.3.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
filelock==3.16.1
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1720359039462/work
fsspec==2025.3.0
geometry-msgs==4.9.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
image-geometry==3.2.1
imageio==2.35.1
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1725921340658/work
interactive-markers==2.3.2
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1695379923772/work
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.12
launch-testing==1.0.10
launch-testing-ros==0.19.12
launch-xml==1.0.10
launch-yaml==1.0.10
lazy_loader==0.4
lifecycle-msgs==1.2.2
logging-demo==0.20.5
map-msgs==2.1.0
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1695076686224/work
message-filters==4.3.8
munkres==1.1.4
nav-msgs==4.9.0
networkx==3.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
opencv-python==4.10.0
opencv-python-headless==4.10.0
osrf-pycommon==2.1.6
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1719903565503/work
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1712242996588/work
protobuf==5.29.5
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1724616129934/work
PyQt5==5.15.9
PyQt5-sip==12.12.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-qt-binding==1.1.2
PyWavelets==1.4.1
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
requests==2.32.4
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2component==0.18.14
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2topic==0.18.14
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
shape-msgs==4.9.0
sip @ file:///home/conda/feedstock_root/build_artifacts/sip_1697300436403/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
teleop-twist-keyboard==2.4.0
tensorboardX==2.6.2.2
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
thop==0.1.1.post2209072238
tifffile==2023.7.10
timm==1.0.22
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1727974628237/work
topic-monitor==0.20.5
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1717722826518/work
tqdm==4.67.1
trajectory-msgs==4.9.0
turtlesim==1.4.2
typing_extensions==4.13.2
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1695847997538/work
unique-identifier-msgs==2.2.1
urllib3==2.2.3
visualization-msgs==4.9.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1731262100163/work
